Refactor and re-record tests; Add properties for NamespacesTopicsSubscriptionsRule sample
