+++ v2/config/samples/servicebus/v1beta/v1beta20210101preview_namespacestopicssubscriptionsrule.yaml
@@ -6,3 +6,12 @@ metadata:
 spec:
   owner:
     name: sampletopicsubscription
+  action:
+    sqlExpression: "SET myproperty=myvalue"
+    requiresPreprocessing: false
+    compatibilityLevel: 20
+  correlationFilter:
+    contentType: text
+  filterType: SqlFilter
+  sqlFilter:
+    sqlExpression: "StoreId IN ('Store1','Store2','Store3')"
