+++ v2/internal/controllers/recordings/Test_Samples_CreationAndDeletion/Test_Servicebus_v1beta_CreationAndDeletion.yaml
@@ -482,13 +482,15 @@ interactions:
     code: 200
     duration: ""
 - request:
-    body: '{"name":"asotestlijqaw"}'
+    body: '{"name":"asotestlijqaw","properties":{"action":{"compatibilityLevel":20,"requiresPreprocessing":false,"sqlExpression":"SET
+      myproperty=myvalue"},"correlationFilter":{"contentType":"text"},"filterType":"SqlFilter","sqlFilter":{"sqlExpression":"StoreId
+      IN (''Store1'',''Store2'',''Store3'')"}}}'
     form: {}
     headers:
       Accept:
       - application/json
       Content-Length:
-      - "24"
+      - "285"
       Content-Type:
       - application/json
       Test-Request-Attempt:
@@ -496,7 +498,9 @@ interactions:
     url: https://management.azure.com/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/asotest-rg-vbehns/providers/Microsoft.ServiceBus/namespaces/asotestpqebnz/topics/asotestyhaxhd/subscriptions/asotestyyutyj/rules/asotestlijqaw?api-version=2021-01-01-preview
     method: PUT
   response:
-    body: '{"id":"/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/asotest-rg-vbehns/providers/Microsoft.ServiceBus/namespaces/asotestpqebnz/topics/asotestyhaxhd/subscriptions/asotestyyutyj/rules/asotestlijqaw","name":"asotestlijqaw","type":"Microsoft.ServiceBus/namespaces/topics/subscriptions/rules","location":"westcentralus","properties":{"action":{},"filterType":"SqlFilter","sqlFilter":{"sqlExpression":"1=1","compatibilityLevel":20}}}'
+    body: '{"id":"/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/asotest-rg-vbehns/providers/Microsoft.ServiceBus/namespaces/asotestpqebnz/topics/asotestyhaxhd/subscriptions/asotestyyutyj/rules/asotestlijqaw","name":"asotestlijqaw","type":"Microsoft.ServiceBus/namespaces/topics/subscriptions/rules","location":"westcentralus","properties":{"action":{"sqlExpression":"SET
+      myproperty=myvalue","compatibilityLevel":20},"filterType":"SqlFilter","sqlFilter":{"sqlExpression":"StoreId
+      IN (''Store1'',''Store2'',''Store3'')","compatibilityLevel":20}}}'
     headers:
       Cache-Control:
       - no-cache
@@ -528,7 +532,9 @@ interactions:
     url: https://management.azure.com/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/asotest-rg-vbehns/providers/Microsoft.ServiceBus/namespaces/asotestpqebnz/topics/asotestyhaxhd/subscriptions/asotestyyutyj/rules/asotestlijqaw?api-version=2021-01-01-preview
     method: GET
   response:
-    body: '{"id":"/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/asotest-rg-vbehns/providers/Microsoft.ServiceBus/namespaces/asotestpqebnz/topics/asotestyhaxhd/subscriptions/asotestyyutyj/rules/asotestlijqaw","name":"asotestlijqaw","type":"Microsoft.ServiceBus/namespaces/topics/subscriptions/rules","location":"westcentralus","properties":{"action":{},"filterType":"SqlFilter","sqlFilter":{"sqlExpression":"1=1","compatibilityLevel":20}}}'
+    body: '{"id":"/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/asotest-rg-vbehns/providers/Microsoft.ServiceBus/namespaces/asotestpqebnz/topics/asotestyhaxhd/subscriptions/asotestyyutyj/rules/asotestlijqaw","name":"asotestlijqaw","type":"Microsoft.ServiceBus/namespaces/topics/subscriptions/rules","location":"westcentralus","properties":{"action":{"sqlExpression":"SET
+      myproperty=myvalue","compatibilityLevel":20,"parameters":{}},"filterType":"SqlFilter","sqlFilter":{"sqlExpression":"StoreId
+      IN (''Store1'',''Store2'',''Store3'')","compatibilityLevel":20,"parameters":{}}}}'
     headers:
       Cache-Control:
       - no-cache
